Move plot logic into DataAssistantResult
