+++ great_expectations/rule_based_profiler/types/__init__.py
@@ -1,3 +1,7 @@
+from great_expectations.rule_based_profiler.types.data_assistant_result.data_assistant_result import (  # isort:skip
+    DataAssistantResult,
+)
+
 from .altair_configuration import ALTAIR_DEFAULT_CONFIGURATION
 
 from .attributes import Attributes  # isort:skip
@@ -30,4 +34,3 @@ from .parameter_container import (  # isort:skip
     get_fully_qualified_parameter_names,
 )
 from .rule_state import RuleState  # isort:skip
-from .data_assistant_result import DataAssistantResult  # isort:skip
