+++ great_expectations/rule_based_profiler/types/data_assistant_result/__init__.py
@@ -1,0 +1,1 @@
+from .data_assistant_result import DataAssistantResult
